change slot index to type int24

+++ contracts/Hyper.sol
@@ -173,7 +173,7 @@ contract Hyper is IHyper, ReentrancyGuard {
     function syncSlot(
         PoolId poolId,
         Slot storage slot,
-        int128 slotIndex,
+        int24 slotIndex,
         Epoch memory readEpoch
     ) internal {
         uint256 epochsPassed = readEpoch.getEpochsPassed(slot.lastUpdatedTimestamp);
@@ -187,7 +187,7 @@ contract Hyper is IHyper, ReentrancyGuard {
             if (slot.pendingLiquidityGross < 0) {
                 slot.liquidityGross -= BrainMath.abs(slot.pendingLiquidityGross);
                 if (slot.liquidityGross == 0) {
-                    (int16 chunk, uint8 bit) = BitMath.getSlotPositionInBitmap(int24(slotIndex));
+                    (int16 chunk, uint8 bit) = BitMath.getSlotPositionInBitmap(slotIndex);
                     bitmaps[poolId][chunk] = BitMath.flip(bitmaps[poolId][chunk], bit);
                     delete slots[slot.id];
                     return;
@@ -309,8 +309,8 @@ contract Hyper is IHyper, ReentrancyGuard {
 
     function updateLiquidity(
         PoolId poolId,
-        int128 lowerSlotIndex,
-        int128 upperSlotIndex,
+        int24 lowerSlotIndex,
+        int24 upperSlotIndex,
         int256 amount
     ) public nonReentrant started {
         if (lowerSlotIndex >= upperSlotIndex) revert IHyper.PositionInvalidRangeError();
@@ -325,7 +325,7 @@ contract Hyper is IHyper, ReentrancyGuard {
         SlotId lowerSlotId;
         Slot storage lowerSlot;
         {
-            lowerSlotId = getSlotId(poolId, int24(lowerSlotIndex));
+            lowerSlotId = getSlotId(poolId, lowerSlotIndex);
             lowerSlot = slots[lowerSlotId];
             syncSlot(poolId, lowerSlot, lowerSlotIndex, epoch);
         }
@@ -333,7 +333,7 @@ contract Hyper is IHyper, ReentrancyGuard {
         SlotId upperSlotId;
         Slot storage upperSlot;
         {
-            upperSlotId = getSlotId(poolId, int24(upperSlotIndex));
+            upperSlotId = getSlotId(poolId, upperSlotIndex);
             upperSlot = slots[upperSlotId];
             syncSlot(poolId, upperSlot, upperSlotIndex, epoch);
         }
@@ -406,7 +406,7 @@ contract Hyper is IHyper, ReentrancyGuard {
             lowerSlot.pendingLiquidityDelta += int256(addAmountLeft);
             if (lowerSlot.liquidityGross == 0) {
                 // flip slot in bitmap
-                (int16 chunk, uint8 bit) = BitMath.getSlotPositionInBitmap(int24(position.lowerSlotIndex));
+                (int16 chunk, uint8 bit) = BitMath.getSlotPositionInBitmap(position.lowerSlotIndex);
                 bitmaps[pool.id][chunk] = BitMath.flip(bitmaps[pool.id][chunk], bit);
                 // initialize per liquidity outside values
                 if (pool.slotIndex >= position.lowerSlotIndex) {
@@ -421,7 +421,7 @@ contract Hyper is IHyper, ReentrancyGuard {
             upperSlot.pendingLiquidityDelta -= int256(addAmountLeft);
             if (upperSlot.liquidityGross == 0) {
                 // flip slot in bitmap
-                (int16 chunk, uint8 bit) = BitMath.getSlotPositionInBitmap(int24(position.upperSlotIndex));
+                (int16 chunk, uint8 bit) = BitMath.getSlotPositionInBitmap(position.upperSlotIndex);
                 bitmaps[pool.id][chunk] = BitMath.flip(bitmaps[pool.id][chunk], bit);
                 // initialize per liquidity outside values
                 if (pool.slotIndex >= position.upperSlotIndex) {
@@ -476,7 +476,7 @@ contract Hyper is IHyper, ReentrancyGuard {
             lowerSlot.liquidityGross -= removedPending;
 
             if (lowerSlot.liquidityGross == 0) {
-                (int16 chunk, uint8 bit) = BitMath.getSlotPositionInBitmap(int24(position.lowerSlotIndex));
+                (int16 chunk, uint8 bit) = BitMath.getSlotPositionInBitmap(position.lowerSlotIndex);
                 bitmaps[pool.id][chunk] = BitMath.flip(bitmaps[pool.id][chunk], bit);
                 delete slots[lowerSlot.id];
             }
@@ -486,7 +486,7 @@ contract Hyper is IHyper, ReentrancyGuard {
             upperSlot.liquidityGross -= removedPending;
 
             if (upperSlot.liquidityGross == 0) {
-                (int16 chunk, uint8 bit) = BitMath.getSlotPositionInBitmap(int24(position.upperSlotIndex));
+                (int16 chunk, uint8 bit) = BitMath.getSlotPositionInBitmap(position.upperSlotIndex);
                 bitmaps[pool.id][chunk] = BitMath.flip(bitmaps[pool.id][chunk], bit);
                 delete slots[upperSlot.id];
             }
@@ -535,14 +535,14 @@ contract Hyper is IHyper, ReentrancyGuard {
         UD60x18 feeTier;
         uint256 remaining;
         uint256 amountOut;
-        int128 slotIndex;
+        int24 slotIndex;
         UD60x18 sqrtPrice;
         uint256 swapLiquidity;
         uint256 maturedLiquidity;
         int256 pendingLiquidity;
         UD60x18 feesPerLiquidity;
         bool nextSlotInitialized;
-        int128 nextSlotIndex;
+        int24 nextSlotIndex;
         UD60x18 nextSqrtPrice;
     }
 
@@ -577,14 +577,14 @@ contract Hyper is IHyper, ReentrancyGuard {
         while (swapDetails.remaining > 0) {
             {
                 // Get the next slot or the border of a bitmap
-                (int16 chunk, uint8 bit) = BitMath.getSlotPositionInBitmap(int24(swapDetails.slotIndex));
+                (int16 chunk, uint8 bit) = BitMath.getSlotPositionInBitmap(swapDetails.slotIndex);
                 (bool hasNextSlot, uint8 nextSlotBit) = BitMath.findNextSlotWithinChunk(
                     bitmaps[poolId][chunk],
                     bit,
                     tokenIn == PoolToken.A ? BitMath.SearchDirection.Left : BitMath.SearchDirection.Right
                 );
                 swapDetails.nextSlotInitialized = hasNextSlot;
-                swapDetails.nextSlotIndex = int128(chunk) * 256 + int8(nextSlotBit);
+                swapDetails.nextSlotIndex = int24(chunk) * 256 + int8(nextSlotBit);
                 swapDetails.nextSqrtPrice = BrainMath.getSqrtPriceAtSlot(swapDetails.nextSlotIndex);
             }
             uint256 remainingFeeAmount = fromUD60x18(swapDetails.feeTier.mul(toUD60x18(swapDetails.remaining)).ceil());
